DB schema update and integration, models update

+++ ServiceDB/DBO/Tables/Vehicles.sql
@@ -5,7 +5,7 @@
 	[model] NVARCHAR(50) NOT NULL,
 	[type_id] INT NOT NULL,
 	[registration] NVARCHAR(50) NOT NULL,
-	[avg_fuel_consumption] FLOAT NOT NULL
+	[avg_fuel_consumption] DECIMAL(18, 1) NOT NULL
 	CONSTRAINT FK_type_id FOREIGN KEY (type_id)
         REFERENCES Vehicle_types (id)
 )
